refactor: flatten cartesian pipeline and remove legacy signatures
