+++ packages/charts/src/cartesian/bar.test.js
@@ -161,144 +161,4 @@ describe("bar", () => {
       expect(barFn.isBar).toBe(true)
     })
   })
-
-  describe("renderXAxis()", () => {
-    it("should return a function marked as isAxis", () => {
-      const result = bar.renderXAxis(entity, {}, api)
-
-      expect(typeof result).toBe("function")
-      expect(result.isAxis).toBe(true)
-    })
-
-    it("should render X axis when called with context", () => {
-      const xAxisFn = bar.renderXAxis(entity, {}, api)
-
-      const context = {
-        entity,
-        xScale: {
-          domain: () => ["Jan", "Feb", "Mar"],
-          range: () => [50, 750],
-          bandwidth: () => 200,
-        },
-        yScale: {
-          domain: () => [0, 200],
-          range: () => [370, 20],
-        },
-        dimensions: {
-          width: 800,
-          height: 400,
-          padding: { top: 20, right: 50, bottom: 30, left: 50 },
-        },
-      }
-
-      // Mock xScale methods
-      context.xScale = vi.fn((cat) => {
-        const map = { Jan: 100, Feb: 350, Mar: 600 }
-        return map[cat] || 0
-      })
-      context.xScale.domain = vi.fn(() => ["Jan", "Feb", "Mar"])
-      context.xScale.range = vi.fn(() => [50, 750])
-      context.xScale.bandwidth = vi.fn(() => 200)
-
-      const result = xAxisFn(context)
-      const container = document.createElement("div")
-      render(result, container)
-
-      // Should render axis line and labels
-      const axisLine = container.querySelector("line")
-      expect(axisLine).toBeTruthy()
-    })
-  })
-
-  describe("renderYAxis()", () => {
-    it("should return a function", () => {
-      const result = bar.renderYAxis(entity, {}, api)
-
-      expect(typeof result).toBe("function")
-    })
-
-    it("should render Y axis when called with context", () => {
-      const yAxisFn = bar.renderYAxis(entity, {}, api)
-
-      const context = {
-        entity,
-        yScale: {
-          domain: () => [0, 200],
-          range: () => [370, 20],
-          ticks: () => [0, 50, 100, 150, 200],
-        },
-        dimensions: {
-          width: 800,
-          height: 400,
-          padding: { top: 20, right: 50, bottom: 30, left: 50 },
-        },
-      }
-
-      context.yScale = vi.fn((val) => 370 - (val / 200) * 350)
-      context.yScale.domain = vi.fn(() => [0, 200])
-      context.yScale.range = vi.fn(() => [370, 20])
-      context.yScale.ticks = vi.fn(() => [0, 50, 100, 150, 200])
-
-      const result = yAxisFn(context)
-      const container = document.createElement("div")
-      render(result, container)
-
-      // Should render axis line
-      const axisLine = container.querySelector("line")
-      expect(axisLine).toBeTruthy()
-    })
-  })
-
-  describe("renderCartesianGrid()", () => {
-    it("should return a function", () => {
-      const result = bar.renderCartesianGrid(entity, {}, api)
-
-      expect(typeof result).toBe("function")
-    })
-
-    it("should render grid when called with context", () => {
-      const gridFn = bar.renderCartesianGrid(entity, {}, api)
-
-      const context = {
-        entity,
-        xScale: {
-          domain: () => ["Jan", "Feb", "Mar"],
-          range: () => [50, 750],
-          bandwidth: () => 200,
-        },
-        yScale: {
-          domain: () => [0, 200],
-          range: () => [370, 20],
-          ticks: () => [0, 50, 100, 150, 200],
-        },
-        dimensions: {
-          width: 800,
-          height: 400,
-          padding: { top: 20, right: 50, bottom: 30, left: 50 },
-        },
-      }
-
-      // Mock scales
-      context.xScale = vi.fn((cat) => {
-        const map = { Jan: 100, Feb: 350, Mar: 600 }
-        return map[cat] || 0
-      })
-      context.xScale.domain = vi.fn(() => ["Jan", "Feb", "Mar"])
-      context.xScale.range = vi.fn(() => [50, 750])
-      context.xScale.bandwidth = vi.fn(() => 200)
-
-      context.yScale = vi.fn((val) => 370 - (val / 200) * 350)
-      context.yScale.domain = vi.fn(() => [0, 200])
-      context.yScale.range = vi.fn(() => [370, 20])
-      context.yScale.ticks = vi.fn(() => [0, 50, 100, 150, 200])
-
-      const result = gridFn(context)
-      const container = document.createElement("div")
-      render(result, container)
-
-      // Should render grid lines
-      const lines = container.querySelectorAll("line")
-      expect(lines.length).toBeGreaterThan(0)
-    })
-  })
 })
