+++ packages/charts/src/cartesian/line.js
@@ -1,21 +1,15 @@
 /* eslint-disable no-magic-numbers */
 import { repeat, svg } from "@inglorious/web"
 
-import { createBrushComponent } from "../component/brush.js"
-import { renderGrid } from "../component/grid.js"
-import { renderLegend } from "../component/legend.js"
-import { createTooltipComponent } from "../component/tooltip.js"
-import { renderXAxis } from "../component/x-axis.js"
-import { renderYAxis } from "../component/y-axis.js"
 import { chart } from "../index.js"
 import { renderDot } from "../shape/dot.js"
 import {
   createBandCenterScale,
+  createCartesianRenderer,
   getResolvedEntity,
   inferSeriesDataKey,
   resolveCategoryLabel,
-} from "../utils/cartesian-helpers.js"
-import { createCartesianRenderer } from "../utils/cartesian-renderer.js"
+} from "../utils/cartesian-utils.js"
 import { PALETTE_DEFAULT } from "../utils/constants.js"
 import { getTransformedData } from "../utils/data-utils.js"
 import { generateLinePath } from "../utils/paths.js"
@@ -43,123 +37,8 @@ export const line = {
       dataKeys,
       originalEntity: { ...entity, data: transformedData },
     }),
+    renderChart: renderComposedChart,
   }),
-
-  /**
-   * Composition rendering entry point for line charts.
-   * @param {import('../types/charts').ChartEntity} entity
-   * @param {{ children: any[]|any, config?: Record<string, any> }} params
-   * @param {import('@inglorious/web').Api} api
-   * @returns {import('lit-html').TemplateResult}
-   */
-  renderLineChart(entity, { children, config = {} }, api) {
-    return renderComposedChart(entity, { children, config }, api)
-  },
-
-  /**
-   * Composition sub-render for cartesian grid.
-   * @param {import('../types/charts').ChartEntity} entity
-   * @param {{ config?: Record<string, any> }} params
-   * @param {import('@inglorious/web').Api} api
-   * @returns {(ctx: Record<string, any>) => import('lit-html').TemplateResult}
-   */
-  renderCartesianGrid(entity, { config = {} }, api) {
-    const gridFn = (ctx) => {
-      const { xScale, yScale, dimensions } = ctx
-      const [start, end] = xScale.domain()
-      const ticks = []
-      for (let i = Math.ceil(start); i <= Math.floor(end); i++) {
-        ticks.push(i)
-      }
-
-      const customXScale = Object.assign(xScale.copy(), { ticks: () => ticks })
-
-      return renderGrid(
-        ctx.entity,
-        {
-          ...dimensions,
-          ...config,
-          yScale,
-          xScale: customXScale,
-          ticks: ticks,
-          customXTicks: ticks,
-        },
-        api,
-      )
-    }
-    gridFn.isGrid = true
-    return gridFn
-  },
-
-  /**
-   * Composition sub-render for X axis.
-   * @param {import('../types/charts').ChartEntity} entity
-   * @param {{ config?: Record<string, any> }} params
-   * @param {import('@inglorious/web').Api} api
-   * @returns {(ctx: Record<string, any>) => import('lit-html').TemplateResult}
-   */
-  renderXAxis(entity, { config = {} }, api) {
-    const axisFn = (ctx) => {
-      const { xScale, yScale, dimensions, entity: currentEntity } = ctx
-      const [viewStart, viewEnd] = xScale.domain()
-
-      const scaleTicks = []
-      for (let i = Math.ceil(viewStart); i <= Math.floor(viewEnd); i++) {
-        scaleTicks.push(i)
-      }
-
-      const labels = scaleTicks.map((tick) => {
-        const dataForLabels =
-          ctx.fullEntity === ctx.entity ? ctx.entity.data : currentEntity.data
-        const item = dataForLabels[tick]
-        return item?.[config.dataKey] || item?.name || item?.x || String(tick)
-      })
-
-      const customXScaleForAxis = Object.assign(xScale.copy(), {
-        ticks: () => scaleTicks,
-      })
-
-      return renderXAxis(
-        { ...currentEntity, xLabels: labels },
-        {
-          ...dimensions,
-          yScale,
-          xScale: customXScaleForAxis,
-          customTicks: scaleTicks,
-          tickValues: scaleTicks,
-          tickFormat: (d) => {
-            const idx = Math.round(d)
-            const labelIdx = scaleTicks.indexOf(idx)
-            return labelIdx !== -1 ? labels[labelIdx] : ""
-          },
-        },
-        api,
-      )
-    }
-    axisFn.isAxis = true
-    return axisFn
-  },
-
-  /**
-   * Composition sub-render for Y axis.
-   * @param {import('../types/charts').ChartEntity} entity
-   * @param {{ config?: Record<string, any> }} params
-   * @param {import('@inglorious/web').Api} api
-   * @returns {(ctx: Record<string, any>) => import('lit-html').TemplateResult}
-   */
-  // eslint-disable-next-line no-unused-vars
-  renderYAxis(entity, { config = {} }, api) {
-    const axisFn = (ctx) => {
-      const ticks = ctx.yScale.ticks ? ctx.yScale.ticks(5) : ctx.yScale.domain()
-      return renderYAxis(
-        ctx.entity,
-        { yScale: ctx.yScale, customTicks: ticks, ...ctx.dimensions },
-        api,
-      )
-    }
-    axisFn.isAxis = true
-    return axisFn
-  },
 
   /**
    * Composition sub-render for line paths.
@@ -299,38 +178,6 @@ export const line = {
     dotsFn.isDots = true
     return dotsFn
   },
-
-  /**
-   * Composition sub-render for legend.
-   * @param {import('../types/charts').ChartEntity} entity
-   * @param {{ config?: Record<string, any> }} params
-   * @param {import('@inglorious/web').Api} api
-   * @returns {(ctx: Record<string, any>) => import('lit-html').TemplateResult}
-   */
-  renderLegend: (entity, { config = {} }, api) => {
-    const legendFn = (ctx) => {
-      const { dataKeys = [], labels = [], colors = [] } = config
-      const series = dataKeys.map((key, i) => ({
-        name: labels[i] || key,
-        color: colors[i % colors.length] || PALETTE_DEFAULT[0],
-      }))
-      return renderLegend(entity, { series, ...ctx.dimensions }, api)
-    }
-    legendFn.isLegend = true
-    return legendFn
-  },
-
-  /**
-   * Composition sub-render for tooltip overlay.
-   * @type {(entity: import('../types/charts').ChartEntity, params: { config?: Record<string, any> }, api: import('@inglorious/web').Api) => (ctx: Record<string, any>) => import('lit-html').TemplateResult}
-   */
-  renderTooltip: createTooltipComponent(),
-
-  /**
-   * Composition sub-render for brush control.
-   * @type {(entity: import('../types/charts').ChartEntity, params: { config?: Record<string, any> }, api: import('@inglorious/web').Api) => (ctx: Record<string, any>) => import('lit-html').TemplateResult}
-   */
-  renderBrush: createBrushComponent(),
 }
 
 function ensureClipPathId(entity) {
